server/requirements.txt
fastapi
uvicorn
sqlmodel
sqlalchemy
asyncpg
bcrypt
pydantic
pydantic-settings
PyJWT
python-jose
cryptography == 39.0.1